filled in dropdown adn restructured pull code into a new file. Every page now pulls all game docs but that it
