+++ routes/game-routes.js
@@ -1,3 +1,4 @@
+const pull = require("./src/pullGameData");
 const express = require("express");
 const gameRouter = express.Router();
 
@@ -5,40 +6,15 @@ const firebaseAdmin = require("firebase-admin");
 const db = firebaseAdmin.firestore();
 
 /**
- * Get the all game data from database. Not including scores.
- * @author: jk
- * @param {string} gameId The id of the game in the database.
- * @returns {Promise<{gameId: *, path: *, playCount: *, name: *}>}
- */
-async function getGame(gameId) {
-    const gameDoc = db.collection("games").doc(gameId);
-    return gameDoc.get()
-        .then(snapshot => {
-            const data = snapshot.data();
-            if (!data)
-                return;  // bad game; return nothing
-            return {
-                name: data.name,
-                shortDesc: data.shortDesc,
-                path: data.path,
-                imgPath: data.imgPath,
-                playCount: data.timesPlayed,
-                gameId: snapshot.id,
-            };
-        })
-        .catch(err => {
-            console.log('Error getting documents', err);
-        });
-}
-
-/**
  * Load the game object from the database to get it's path. Then render the game page
  * @author: jk
  */
 gameRouter.get("/:gameId", async function(req, res) {
-    const gameContext = await getGame(req.params.gameId);
-    if (gameContext) {
-        res.render(gameContext.path, gameContext);
+    const gameData = await pull.getGames();
+    const dropdownData = pull.getGameDropdownData(gameData);
+    const game = gameData.find(game => game.gameId === req.params.gameId);
+    if (game) {
+        res.render(game.path, {gameId: game.gameId, scoresGameDropdownData: dropdownData});
     } else {
         res.render("not-found", {msg: "We apologise. This game does not exist."});
     }
